# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Warehouse Products API",
    "version": "1.0.0",
    "description": "API for managing products in a warehouse, with user authentication using JWT."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local API server"
    }
  ],
  "tags": [
    { "name": "Auth", "description": "Endpoints related to user authentication" },
    { "name": "Products", "description": "Endpoints related to product management" },
    { "name": "Public", "description": "Endpoints that do NOT require authentication" },
    { "name": "Protected", "description": "Endpoints that require authentication" }
  ],

  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "Create a new user with username and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterUser" }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered successfully" },
          "400": { "description": "User already exists" }
        }
      }
    },

    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "description": "Login with username and password to receive a JWT access token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginUser" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "400": { "description": "Invalid credentials" }
        }
      }
    },

    "/products": {
      "get": {
        "tags": ["Public", "Products"],
        "summary": "Get all products with pagination and filters",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1, "minimum": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10, "minimum": 1 } },
          { "name": "name", "in": "query", "schema": { "type": "string" } },
          { "name": "minPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalProducts": { "type": "integer" },
                    "page": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "totalPages": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Product" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid pagination or filter values" }
        }
      },

      "post": {
        "tags": ["Products", "Protected"],
        "summary": "Create a new product",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewProduct" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },

      "delete": {
        "tags": ["Products", "Protected"],
        "summary": "Delete multiple products",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteProducts" }
            }
          }
        },
        "responses": {
          "204": { "description": "Products deleted" },
          "401": { "description": "Unauthorized" }
        }
      }
    },

    "/products/{id}": {
      "get": {
        "tags": ["Public", "Products"],
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Product" } }
            }
          },
          "404": { "description": "Product not found" }
        }
      },

      "put": {
        "tags": ["Products", "Protected"],
        "summary": "Update product",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/NewProduct" } }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Product" } }
            }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Product not found" }
        }
      },

      "delete": {
        "tags": ["Products", "Protected"],
        "summary": "Delete product by ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Product deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Product not found" }
        }
      }
    }
  },

  "components": {
    "schemas": {
      "RegisterUser": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },

      "LoginUser": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },

      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "JWT token" }
        }
      },

      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "price": { "type": "number" },
          "quantity": { "type": "integer" }
        }
      },

      "NewProduct": {
        "type": "object",
        "required": ["name", "price", "quantity"],
        "properties": {
          "name": { "type": "string" },
          "price": { "type": "number" },
          "quantity": { "type": "integer" }
        }
      },

      "DeleteProducts": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" }
          }
        }
      }
    },

    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}